UAR-369: Refactored controller to prevent code reuse
